refactor: use custom hooks for data fetching
